#679 Replace explicit `Thread` creation with managed service executor and refactor `CommandListener`
